Implemented JDABoot.Settings and ConfigProvider
With the ConfigProvider the user of JDABoot can specify some code on his own and it is possible to not use the default config.properties provided.
The Settings were implemented to future-proof JDABoot and prevent the constructor and run() methods from getting messy because of too many parameters that are passed to them.

**I strongly recommend to remove the newly deprecated run() method as soon as possible.**
