+++ .idea/compiler.xml
@@ -7,7 +7,7 @@
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.28/a2ff5da8bcd8b1b26f36b806ced63213362c6dcc/lombok-1.18.28.jar" />
         </processorPath>
-        <module name="de.jonafaust.jda-boot.main" />
+        <module name="jda-boot.main" />
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel target="17" />
