+++ .idea/modules.xml
@@ -2,7 +2,6 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/de.jonafaust.jda-boot.main.iml" filepath="$PROJECT_DIR$/.idea/modules/de.jonafaust.jda-boot.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/jda-boot.main.iml" filepath="$PROJECT_DIR$/.idea/modules/jda-boot.main.iml" />
     </modules>
   </component>
